src/requirements.txt
fastapi
uvicorn
mangum
requests
python-dotenv
